+++ apps/mail/mail/rootfs/etc/cont-init.d/00-hosts.sh
@@ -6,7 +6,7 @@ set -o nounset
 set -o pipefail
 
 declare -a HOSTS_TO_CHECK=(
-"antivirus"
+#"antivirus"
 "${MP_DATABASE_HOST}"
 #"fail2ban"
 "opendkim"
@@ -23,9 +23,9 @@ for HOST_TO_CHECK in "${HOSTS_TO_CHECK[@]}"; do
       echo "[INFO] Container IP found, adding a new record in /etc/hosts"
       echo "${IP} ${HOST_TO_CHECK}" >> /etc/hosts
     else
-      echo "[ERROR] Container IP not found with embedded DNS server... Abort !"
-      echo "[ERROR] Check your ${HOST_TO_CHECK} environment variable"
-      exit 1
+      echo "[WARNING] Container IP not found with embedded DNS server... Abort!"
+      echo "[WARNING] Check your ${HOST_TO_CHECK} environment variable"
+      #exit 1
     fi
   else
     echo "[INFO] ${HOST_TO_CHECK} hostname found in /etc/hosts"