+++ tools/prepare_folders.sh
@@ -8,10 +8,10 @@ set -o pipefail
 function prepare_folders {
   echo_ok "Setting up folder structure"
 
-  local DATA_DIR="${1}"
-  local LOG_DIR="${2}"
+  local data_dir="${1}"
+  local log_dir="${2}"
 
-  local -a DATA_SUBDIRS=(
+  local -a data_subdirs=(
     'acme.sh'
     'antivirus'
     'certs'
@@ -26,21 +26,21 @@ function prepare_folders {
     'unbound'
   )
 
-  local -a LOG_SUBDIRS=(
+  local -a log_subdirs=(
     'antivirus'
     'mail'
     'nginx'
     'roundcubemail'
   )
 
-  check_or_create_dir_or_exit "${DATA_DIR}"
-  for SUBDIR in "${DATA_SUBDIRS[@]}"; do
-    check_or_create_dir_or_exit "${DATA_DIR}/${SUBDIR}"
+  check_or_create_dir_or_exit "${data_dir}"
+  for SUBDIR in "${data_subdirs[@]}"; do
+    check_or_create_dir_or_exit "${data_dir}/${SUBDIR}"
   done
 
-  check_or_create_dir_or_exit "${LOG_DIR}"
-  for SUBDIR in "${LOG_SUBDIRS[@]}"; do
-    check_or_create_dir_or_exit "${LOG_DIR}/${SUBDIR}"
+  check_or_create_dir_or_exit "${log_dir}"
+  for SUBDIR in "${log_subdirs[@]}"; do
+    check_or_create_dir_or_exit "${log_dir}/${SUBDIR}"
   done
 
   echo_ok_verbose "Folder structure created successfully"